+++ source_files/QT_engine.pri
@@ -159,7 +159,7 @@ HEADERS += engine/sim/plot.hpp
 HEADERS += engine/sim/proc.hpp
 HEADERS += engine/sim/proc_rng.hpp
 HEADERS += engine/sim/profileset.hpp
-HEADERS += engine/sim/profileset_controller.hpp
+HEADERS += engine/sim/profileset_control.hpp
 HEADERS += engine/sim/progress_bar.hpp
 HEADERS += engine/sim/raid_event.hpp
 HEADERS += engine/sim/reforge_plot.hpp
@@ -351,7 +351,7 @@ SOURCES += engine/sim/plot.cpp
 SOURCES += engine/sim/proc.cpp
 SOURCES += engine/sim/proc_rng.cpp
 SOURCES += engine/sim/profileset.cpp
-SOURCES += engine/sim/profileset_controller.cpp
+SOURCES += engine/sim/profileset_control.cpp
 SOURCES += engine/sim/progress_bar.cpp
 SOURCES += engine/sim/raid_event.cpp
 SOURCES += engine/sim/reforge_plot.cpp
